revise type of 'exit_reason' and 'inst_len' in vcpu_arch
 SDM 24.9.1 Volume3:
 - 'Exit reason' field in VMCS is 32 bits.

 SDM 24.9.4 in Volume3
 - 'VM-exit instruction length' field
   in VMCS is 32 bits.

 This patch is to redefine the data types of above fields
 in 'struct vcpu_arch' and udpate the code using these
 two fields.

+++ include/arch/x86/guest/vcpu.h
@@ -214,10 +214,10 @@ struct vcpu_arch {
 	uint64_t msr_tsc_aux;
 
 	/* VCPU context state information */
-	uint64_t exit_reason;
+	uint32_t exit_reason;
 	uint64_t exit_interrupt_info;
 	uint64_t exit_qualification;
-	uint8_t inst_len;
+	uint32_t inst_len;
 
 	/* Information related to secondary / AP VCPU start-up */
 	uint8_t cpu_mode;
